Move NLMTA lookup out to a service
